# Use BVHNavigator for TestEm3