--- a/examples/TestEm3/CMakeLists.txt
+++ b/examples/TestEm3/CMakeLists.txt
@@ -7,7 +7,7 @@ if(NOT TARGET G4HepEm::g4HepEm)
 endif()
 
 # TestEm3 for validation of particle transportation with GPUs.
-add_executable(TestEm3 TestEm3.cpp TestEm3.cu electrons.cu gammas.cu relocation.cu)
+add_executable(TestEm3 TestEm3.cpp TestEm3.cu electrons.cu gammas.cu)
 target_link_libraries(TestEm3 PRIVATE AdePT CopCore::CopCore VecGeom::vecgeom VecGeom::vecgeomcuda_static VecGeom::vgdml ${Geant4_LIBRARIES} G4HepEm::g4HepEmData G4HepEm::g4HepEmInit G4HepEm::g4HepEmRun)
 set_target_properties(TestEm3 PROPERTIES CUDA_SEPARABLE_COMPILATION ON CUDA_RESOLVE_DEVICE_SYMBOLS ON)