… examples/TestEm3/relocation.cu