TestEm3: Integrate relocation

For this very simple geometry, it's faster to not relocate in a different kernel, even with the `LoopNavigator`.

For PbWO4 + lAr, time to simulate 10,000 particles:
```
no field: 33.53s -> 21.07s (-37%)
B_z = 1T: 59.21s -> 46.77s (-21%)
```

---

TestEm3: Switch to BVHNavigator

This is slightly slower than the `LoopNavigator` because the geometry only contains boxes, but it's better to match Example 11 and not optimize for a particular setup.

For PbWO4 + lAr, time to simulate 10,000 particles:
```
no field: 21.07s -> 22.35s (+6%)
B_z = 1T: 46.77s -> 51.41s (+10%)
```